# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Job Portal API",
        "description": "API для управління вакансіями",
        "version": "1.0.0"
    },
    "paths": {
        "/api/vacancies": {
            "get": {
                "tags": [
                    "Vacancies"
                ],
                "summary": "Отримати список вакансій",
                "operationId": "251012acba6eed6255a066871efbfb30",
                "parameters": [
                    {
                        "name": "title",
                        "in": "query",
                        "description": "Пошук за назвою вакансії",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "department_id",
                        "in": "query",
                        "description": "Фільтрація за ID департаменту",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "min_salary",
                        "in": "query",
                        "description": "Мінімальна зарплата",
                        "required": false,
                        "schema": {
                            "type": "number"
                        }
                    },
                    {
                        "name": "max_salary",
                        "in": "query",
                        "description": "Максимальна зарплата",
                        "required": false,
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Список вакансій",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Vacancy"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Vacancies"
                ],
                "summary": "Створити нову вакансію",
                "operationId": "fb3b526d24d244e103806f8df7d1cb61",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/VacancyRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Вакансія успішно створена",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Vacancy"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "VacancyRequest": {
                "required": [
                    "title",
                    "description"
                ],
                "properties": {
                    "title": {
                        "description": "Title of the vacancy",
                        "type": "string"
                    },
                    "description": {
                        "description": "Description of the vacancy",
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "Vacancy": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "title": {
                        "type": "string",
                        "example": "PHP Developer"
                    },
                    "description": {
                        "type": "string",
                        "example": "We are looking for a PHP developer with Laravel experience."
                    },
                    "min_salary": {
                        "type": "number",
                        "example": 1000
                    },
                    "max_salary": {
                        "type": "number",
                        "example": 2000
                    },
                    "department_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "is_hot": {
                        "type": "boolean",
                        "example": true
                    }
                },
                "type": "object"
            }
        }
    },
    "tags": [
        {
            "name": "Vacancies",
            "description": "Vacancies"
        }
    ]
}